shuffle things around and add file filter
