+++ src/scopes/activityReport/reportText.js
@@ -49,68 +49,6 @@ const objectiveTtaProvidedPosNeg = (pos = true) => {
   WHERE "ActivityReportObjectives"."ttaProvided"${a}`;
 };
 
-const activityReportResourcePosNeg = (pos = true) => {
-  const a = pos ? '' : ' IS NULL OR "Resources"."url"';
-
-  return `
-  SELECT DISTINCT
-    "ActivityReports"."id"
-  FROM "ActivityReports"
-  LEFT JOIN "ActivityReportResources"
-  ON "ActivityReportResources"."activityReportId" = "ActivityReports"."id"
-  LEFT JOIN "Resources"
-  ON "Resources"."id" = "ActivityReportResources"."resourceId"
-  WHERE "Resources"."url"${a}`;
-};
-
-const activityReportGoalResourcePosNeg = (pos = true) => {
-  const a = pos ? '' : ' IS NULL OR "Resources"."url"';
-
-  return `
-  SELECT DISTINCT
-    "ActivityReports"."id"
-  FROM "ActivityReports"
-  LEFT JOIN "ActivityReportGoals"
-  ON "ActivityReportGoals"."activityReportId" = "ActivityReports"."id"
-  LEFT JOIN "ActivityReportGoalResources"
-  ON "ActivityReportGoalResources"."activityReportGoalId" = "ActivityReportGoals"."id"
-  LEFT JOIN "Resources"
-  ON "Resources"."id" = "ActivityReportGoalResources"."resourceId"
-  WHERE "Resources"."url"${a}`;
-};
-
-const activityReportObjectiveResourcePosNeg = (pos = true) => {
-  const a = pos ? '' : ' IS NULL OR "Resources"."url"';
-
-  return `
-  SELECT DISTINCT
-    "ActivityReports"."id"
-  FROM "ActivityReports"
-  LEFT JOIN "ActivityReportObjectives"
-  ON "ActivityReportObjectives"."activityReportId" = "ActivityReports"."id"
-  LEFT JOIN "ActivityReportObjectiveResources"
-  ON "ActivityReportObjectiveResources"."activityReportObjectiveId" = "ActivityReportObjectives"."id"
-  LEFT JOIN "Resources"
-  ON "Resources"."id" = "ActivityReportObjectiveResources"."resourceId"
-  WHERE "Resources"."url"${a}`;
-};
-
-const nextStepsResourcePosNeg = (pos = true) => {
-  const a = pos ? '' : ' IS NULL OR "Resources"."url"';
-
-  return `
-  SELECT DISTINCT
-    "ActivityReports"."id"
-  FROM "ActivityReports"
-  LEFT JOIN "NextSteps"
-  ON "NextSteps"."activityReportId" = "ActivityReports"."id"
-  LEFT JOIN "NextStepResources"
-  ON "NextSteps"."id" = "NextStepResources"."nextStepId"
-  LEFT JOIN "Resources"
-  ON "Resources"."id" = "NextStepResources"."resourceId"
-  WHERE "Resources"."url"${a}`;
-};
-
 const activityReportContextPosNeg = (pos = true) => {
   const a = pos ? '' : ' IS NULL OR "ActivityReports"."context"';
 
@@ -140,10 +78,6 @@ export function withReportText(searchText) {
       filterAssociation(argsPosNeg(true), search, false, 'ILIKE'),
       filterAssociation(objectiveTitlePosNeg(true), search, false, 'ILIKE'),
       filterAssociation(objectiveTtaProvidedPosNeg(true), search, false, 'ILIKE'),
-      filterAssociation(activityReportResourcePosNeg(true), search, false, 'ILIKE'),
-      filterAssociation(activityReportGoalResourcePosNeg(true), search, false, 'ILIKE'),
-      filterAssociation(activityReportObjectiveResourcePosNeg(true), search, false, 'ILIKE'),
-      filterAssociation(nextStepsResourcePosNeg(true), search, false, 'ILIKE'),
       filterAssociation(activityReportContextPosNeg(true), search, false, 'ILIKE'),
       filterAssociation(additionalNotesPosNeg(true), search, false, 'ILIKE'),
     ],
@@ -159,10 +93,6 @@ export function withoutReportText(searchText) {
       filterAssociation(argsPosNeg(false), search, false, 'NOT ILIKE'),
       filterAssociation(objectiveTitlePosNeg(false), search, false, 'NOT ILIKE'),
       filterAssociation(objectiveTtaProvidedPosNeg(false), search, false, 'NOT ILIKE'),
-      filterAssociation(activityReportResourcePosNeg(false), search, false, 'NOT ILIKE'),
-      filterAssociation(activityReportGoalResourcePosNeg(false), search, false, 'NOT ILIKE'),
-      filterAssociation(activityReportObjectiveResourcePosNeg(false), search, false, 'NOT ILIKE'),
-      filterAssociation(nextStepsResourcePosNeg(false), search, false, 'NOT ILIKE'),
       filterAssociation(activityReportContextPosNeg(false), search, false, 'NOT ILIKE'),
       filterAssociation(additionalNotesPosNeg(false), search, false, 'NOT ILIKE'),
     ],
